Starting work on the gene management for a gene panel. - Fixed redirection after a gene is created to use an HTTP redirect. - After a gene is created, it now redirects to the gene management page. - Fixed transcriptid field of genepanel2genes table to match with the new size of the transcript's ID. - Minor fixes in the code's comments.

+++ src/install/inc-sql-tables.php
@@ -868,7 +868,7 @@ $aTableSQL =
     'CREATE TABLE ' . TABLE_GP2GENE . ' (
     genepanelid SMALLINT(5) UNSIGNED ZEROFILL NOT NULL,
     geneid VARCHAR(25) NOT NULL,
-    transcriptid SMALLINT(5) UNSIGNED ZEROFILL,
+    transcriptid MEDIUMINT(8) UNSIGNED ZEROFILL,
     inheritance VARCHAR(50) NOT NULL,
     id_omim INT(10) UNSIGNED,
     pmid INT(10) UNSIGNED,
@@ -893,7 +893,7 @@ $aTableSQL =
     'CREATE TABLE ' . TABLE_GP2GENE_REV . ' (
     genepanelid SMALLINT(5) UNSIGNED ZEROFILL NOT NULL,
     geneid VARCHAR(25) NOT NULL,
-    transcriptid SMALLINT(5) UNSIGNED ZEROFILL,
+    transcriptid MEDIUMINT(8) UNSIGNED ZEROFILL,
     inheritance VARCHAR(50) NOT NULL,
     id_omim INT(10) UNSIGNED,
     pmid INT(10) UNSIGNED,
